requirements.txt
lask
opencv-python
numpy
livereload